+++ pyslvs_ui/io/yaml.py
@@ -11,6 +11,7 @@ __email__ = "pyslvs@gmail.com"
 
 from re import sub
 from yaml import safe_load, safe_dump
+from qtpy.QtWidgets import QMessageBox
 from .format_editor import FormatEditor
 
 
@@ -29,7 +30,11 @@ class YamlEditor(FormatEditor):
             flow_style = True
         else:
             raise ValueError(f"unsupported option: {self.prefer.file_type_option}")
-        yaml_script = safe_dump(data, default_flow_style=flow_style)
+        try:
+            yaml_script = safe_dump(data, default_flow_style=flow_style)
+        except Exception as e:
+            QMessageBox.warning(self._parent, "Save error", f"{e}")
+            return
         if self.prefer.file_type_option == 1:
             yaml_script = sub(r"\s\s+", " ", yaml_script)
         with open(file_name, 'w+', encoding='utf-8') as f:
@@ -39,4 +44,9 @@ class YamlEditor(FormatEditor):
         """Load YAML file."""
         with open(file_name, 'r', encoding='utf-8') as f:
             yaml_script = f.read()
-        self.load_data(file_name, safe_load(yaml_script))
+        try:
+            data = safe_load(yaml_script)
+        except Exception as e:
+            QMessageBox.warning(self._parent, "Loader error", f"{e}")
+            return
+        self.load_data(file_name, data)