+++ pyslvs_ui/widgets/main_abc.py
@@ -32,115 +32,116 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
 
     @abstractmethod
     def command_reload(self, index: int) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def new_point(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def add_normal_point(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def add_fixed_point(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def edit_point(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def delete_selected_points(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def lock_points(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def new_link(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def edit_link(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def delete_selected_links(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def delete_empty_links(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def constrain_link(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def release_ground(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def add_target_point(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
-    def set_free_move(self, args: Sequence[Tuple[int, Tuple[float, float, float]]]) -> None:
-        ...
+    def set_free_move(self, args: Sequence[
+        Tuple[int, Tuple[float, float, float]]]) -> None:
+        raise NotImplementedError
 
     @abstractmethod
     def add_point_by_pos(self, x: float, y: float) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def set_mouse_pos(self, x: float, y: float) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def get_back_position(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def solve(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def resolve(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def enable_mechanism_actions(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def clone_point(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def copy_coord(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def copy_points_table(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def copy_links_table(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def canvas_context_menu(self, point: QPoint) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def point_context_menu(self, point: QPoint) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def link_context_menu(self, point: QPoint) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def preview_path(
@@ -149,31 +150,31 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
         slider_auto_preview: Dict[int, List[_Coord]],
         vpoints: Sequence[VPoint]
     ) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def reload_canvas(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def output_to(self, format_name: str, format_choose: Sequence[str]) -> str:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def right_input(self) -> bool:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def set_coords_as_current(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def dof(self) -> int:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def save_reply_box(self, title: str, file_name: str) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def input_from(
@@ -181,7 +182,7 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
         format_name: str,
         format_choose: Sequence[str]
     ) -> str:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def input_from_multiple(
@@ -189,7 +190,7 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
         format_name: str,
         format_choose: Sequence[str]
     ) -> List[str]:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def get_graph(self) -> Tuple[
@@ -200,54 +201,54 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
         Mapping[int, int],
         Mapping[int, int]
     ]:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def get_configure(self) -> Mapping[str, Any]:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def project_no_save(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def project_saved(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def merge_result(self, expr: str, path: Sequence[Sequence[_Coord]]) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def check_file_changed(self) -> bool:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def get_storage(self) -> Mapping[str, str]:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def add_empty_links(self, link_color: Mapping[str, str]) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def parse_expression(self, expr: str) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def add_multiple_storage(self, exprs: Mapping[str, str]) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def clear(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def add_points(
         self,
         p_attr: Sequence[Tuple[float, float, str, str, int, float]]
     ) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def add_points_by_graph(
@@ -256,56 +257,56 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
         pos: Dict[int, Tuple[float, float]],
         ground_link: Optional[int]
     ) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def set_window_title_full_path(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def import_pmks_url(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def save_picture_clipboard(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def show_expr(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def py_script(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def export_dxf(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def export_slvs(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def save_pmks(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def export_image(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def show_overview(self) -> None:
-        ...
+        raise NotImplementedError
 
     @staticmethod
     @abstractmethod
     def new() -> MainWindowABC:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def point_alignment(self) -> None:
-        ...
+        raise NotImplementedError
 
     @abstractmethod
     def get_expression(
@@ -313,4 +314,4 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
         points: Optional[Sequence[VPoint]] = None,
         indent: int = -1
     ) -> str:
-        ...
+        raise NotImplementedError